requirements.txt
asciitree==0.3.3
astroid==2.9.3
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
cloudpickle==2.0.0
cycler==0.11.0
dask==2022.2.1
dnspython==2.2.0
docopt==0.6.2
elementpath==2.4.0
email-validator==1.1.3
fasteners==0.17.3
flake8==4.0.1
fonttools==4.29.1
fsspec==2022.2.0
idna==3.3
imageio==2.16.1
isort==5.10.1
kiwisolver==1.3.2
lazy-object-proxy==1.7.1
locket==0.2.1
lxml==4.8.0
matplotlib==3.5.1
mccabe==0.6.1
networkx==2.7.1
numcodecs==0.9.1
numpy==1.22.2
ome-types==0.2.10
opencv-python==4.5.5.62
packaging==21.3
partd==1.2.0
Pillow==9.0.1
Pint==0.18
pipreqs==0.4.11
platformdirs==2.5.1
pycocotools==2.0.4
pycodestyle==2.8.0
pydantic==1.9.0
pyflakes==2.4.0
pylint==2.12.2
pyparsing==3.0.7
python-dateutil==2.8.2
PyWavelets==1.2.0
PyYAML==6.0
requests==2.27.1
scikit-image==0.19.2
scipy==1.8.0
# Editable install with no version control (src==0.1.0)
-e /zhome/e2/e/154260/cell
tifffile==2022.2.9
toml==0.10.2
toolz==0.11.2
torch==1.10.2
torch-summary==1.4.5
torchvision==0.11.3
typing_extensions==4.1.1
urllib3==1.26.8
wrapt==1.13.3
xmlschema==1.9.2
yarg==0.1.9
zarr==2.11.0
